#378: Unfinishend refactoring with unresolvable symbol.
